feat(*): added i18n and merge from main
