+++ src/components/sections/About.astro
@@ -3,7 +3,7 @@ import Section from "../Section.astro";
 
 import { basics } from "@app/resources/cv.json";
 //i18n
-import { useTranslations } from '@app/i18n/utils'
+import { useTranslations } from '@app/utils/i18n'
 
 const { summary } = basics
