+++ src/components/sections/Certificates.astro
@@ -5,7 +5,7 @@ import CertificatesArticle from "@app/components/CertificatesArticle.astro";
 // json cv
 import { certificates } from "@app/resources/cv.json";
 // i18n
-import { useTranslations } from '@app/i18n/utils'
+import { useTranslations } from '@app/utils/i18n'
 
 const t = useTranslations('en')
 ---
