+++ src/components/sections/Education.astro
@@ -4,7 +4,7 @@ import Article from "@app/components/ExperienceArticle.astro";
 import  CV from "@cv";
 import { obtenerFechaActualISO, MONTHS } from "@app/utils/DateFormat";
 // i18n
-import { useTranslations } from '@app/i18n/utils'
+import { useTranslations } from '@app/utils/i18n'
 
 const { education } = CV
 const t = useTranslations('en')
