--- src/i18n/utils.ts
@@ -1,13 +1,0 @@
-import { ui, defaultLang } from './ui';
-
-export function getLangFromUrl(url: URL) {
-  const [, lang] = url.pathname.split('/');
-  if (lang in ui) return lang as keyof typeof ui;
-  return defaultLang;
-}
-
-export function useTranslations(lang : keyof typeof ui = 'en') {
-  return function t(key: keyof typeof ui[typeof defaultLang]) {
-    return ui[lang][key] || ui[defaultLang][key];
-  }
-}
